adding page for timeout error to be used by ngnix
creating help request form and submit it as part of 502

Reverting changes in static page controller and issue post request directly to help_request controller

Adding static pages action for 502

Instead of a true static page, the /502(.html) route is sent to the
StaticPages#timeout_error. This allows for the generation of a
form with the CSRF token.

Resolves ndlib/curate#29 without a need to merge this pull request.

@TODO - This means that the HelpRequest model needs adjustment within
the local instance of CurateND. The 502 error may not have a current
user. Which means we need a fallback email to be the sender of the
help request.

Let user to create help_request on timeout with curate help email when sender email is not available