Integrating relationship reindexer
I would have loved to fire the relationship indexer as part of a
callback (well, at least as part of the model). But that was not
feasible given that the after_save callback fired 5 times during a
single web transaction.

I'd prefer to not piggy back on that behavior. Instead, I've added the
relationship reindexer as part of the curation concern's actor.