Override method "enforce_show_permissions"
Overriding method from Hydra::AccessControlsEnforcement.
The code is identical but moving it seemed to solve a caching issue with
Hydra::PermissionsCache causing a user to not have permission to view
their newly created work, and requiring a restart of the rails server
to reset it. May need to override in catalog_controller as well, but
only doing it here for now.