Adding cap "db:bootstrap" task (CURATE-130)
The three rake tasks will:

-  Optionally create the database for the given environment; If the
   database exists, then this task is skipped.
-  Destructively load the db/schema.rb file into the environment's
   database.  Any data in the database will be lost.
-  Load any associated database seeds for this environment (see
   `./db/seeds/data.rb`).

I believe this will address CURATE-130.

Note: None of this directly impacts the Fedora instance.  However, there
exists a relationship between users in the database and people in
Fedora.  Any Users before the script runs will be deleted, but their
People record in Fedora will remain (and in essence be abandoned).